[chore] rename auth folder & package name to be security
